ParseOkHttpClient becomes ParseHttpClient
