+++ Parse/src/test/java/com/parse/ParseHttpClientTest.java
@@ -47,12 +47,12 @@ public class ParseHttpClientTest {
   @Test
   public void testParseOkHttpClientExecuteWithSuccessResponse() throws Exception {
     doSingleParseHttpClientExecuteWithResponse(
-        200, "OK", "Success", ParseOkHttpClient.createClient(new OkHttpClient.Builder()));  }
+        200, "OK", "Success", ParseHttpClient.createClient(new OkHttpClient.Builder()));  }
 
   @Test
   public void testParseOkHttpClientExecuteWithErrorResponse() throws Exception {
     doSingleParseHttpClientExecuteWithResponse(
-        404, "NOT FOUND", "Error", ParseOkHttpClient.createClient(new OkHttpClient.Builder()));  }
+        404, "NOT FOUND", "Error", ParseHttpClient.createClient(new OkHttpClient.Builder()));  }
 
   // TODO(mengyan): Add testParseURLConnectionHttpClientExecuteWithGzipResponse, right now we can
   // not do that since in unit test env, URLConnection does not use OKHttp internally, so there is
@@ -61,7 +61,7 @@ public class ParseHttpClientTest {
   @Test
   public void testParseOkHttpClientExecuteWithGzipResponse() throws Exception {
     doSingleParseHttpClientExecuteWithGzipResponse(
-        200, "OK", "Success", ParseOkHttpClient.createClient(new OkHttpClient.Builder()));
+        200, "OK", "Success", ParseHttpClient.createClient(new OkHttpClient.Builder()));
   }
 
   private void doSingleParseHttpClientExecuteWithResponse(int responseCode, String responseStatus,
